Added new option get_mapping
When using the Splunk python SDK, the first event returned seems to be used
for selecting the fields for the returned CSV. If some other elasticsearch
document later in the results have some additional fields, these fields are not
included in the results.

This command queries elasticsearch for all fields in the index/indices and
adds any missing fields with a value of null to all events.